(fix|chore) Flytt businesslogikk inn i skjemahook, fiks checkbox-onchange
